requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Faker==24.11.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-openapi3==3.1.0
Flask-RESTful==0.3.10
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.3
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
paginate==0.5.6
paginate-sqlalchemy==0.3.1
pillow==10.3.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.29
structlog==24.1.0
typing_extensions==4.11.0
Werkzeug==3.0.2
WTForms==3.1.2
